fix: Message와 Util 분리
