+++ src/main/java/racingcar/GameMessage.java
@@ -1,0 +1,10 @@
+package racingcar;
+
+public class GameMessage {
+
+    public final String inputNameMessage = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+    public final String inputTimeMessage = "시도할 횟수는 몇회인가요?";
+    public final String executeMessage = "실행 결과";
+    public final String finalWinnerMessage = "최종 우승자";
+
+}
